Address Mod->ModSystem recipe hook move.
